requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# OpenAI and LLM
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
llama-index==0.9.0

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.18
pgvector==0.2.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sqlite3

# Google Services
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Salesforce
simple-salesforce==1.12.5

# Email
smtplib
email-validator==2.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
jinja2==3.1.2

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# Arabic Text Processing
python-bidi==0.4.2
arabic-reshaper==3.0.0
